pyproject.toml
[project]
name = "odoo-bringout-mrp_production_import_xlsx"
version = "16.0.0"
description = "MRP Production Import Components from XLSX - Generate Manufacturing Order from Components in XLSX file"
authors = [
    { name = "Ernad Husremovic", email = "hernad@bring.out.ba" }
]
dependencies = [
    "odoo-bringout-oca-ocb-mrp>=16.0.0",
    "requests>=2.25.1"
]
readme = "README.md"
requires-python = ">= 3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

[project.urls]
homepage = "https://github.com/bringout/0"
repository = "https://github.com/bringout/0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["mrp_production_import_xlsx"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.4.1",
]
